Updating session url and test cases
